superbase-sink/requirements.txt
./quixstreams-2.3.2-py3-none-any.whl
python-dotenv
supabase